fixed text draw, set GCodePoint to use floats, not ints

+++ src/GCodePoint.h
@@ -10,14 +10,15 @@
 
 class GCodePoint{
 public:
-    //TODO: these should be floats
-    int x;
-    int y;
-    int speed;
-    int pressure;
+    
+    float x;
+    float y;
+    float speed;
+    float pressure;
     
     inline GCodePoint( float _x, float _y, float _speed, float _pressure ):x(_x), y(_y), speed(_speed), pressure(_pressure) {}
     
+    //TODO: MAKE SURE THIS WORKS NOW THAT YOU TURNED IT TO FLOATS
     bool equals(GCodePoint other){
         return x==other.x && y==other.y && speed==other.speed && pressure==other.pressure;
     }
